Added "rising tide" and related functions
Currently, score_to_memberships uses a single somewhat-arbitrary threshold on the score function to decide which vertices to actually add.

As we discussed, an alternative is to compute a global score function for each community and then choose the threshold to maximize this score. The function "rising_tide" is essentially a plug-in replacement for score_to_memberships that does this, using (one definition of) conductance as its global score and connectivity as its global requirement.

This commit also adds the various functions required to make rising_tide work, such as a somewhat-fast way to compute conductance for a sequence of sets that doesn't require recomputing the conductance from scratch each time.